Device wrapper - general restructure: Drop pointers to implementation for the iio_device wrappers - DeviceOut, DeviceIn, DeviceGeneric.
These classes should not be exposed by the library. Their only use
is to interface between the iio_devices and the tools based on them.
They provide basic functionality for all the iio_device features.
We can drop the Pimpl idea for these classes if we don't expose them.
Also, these classes need to be moved outside of the official "include" dir.

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>